Switch to async/await
This turned out be a mechanical transformation, to the somewhat
advanced use of streams, and is not totally faithful; in particular,
the processing of input is now serialized; this should fixed by a
follow-up commit soon.

The code is now more straightforward, and should now quite a bit
easier to follow.